+++ framework/src/test/java/com/github/dbmdz/flusswerk/framework/engine/EngineTest.java
@@ -1,5 +1,7 @@
 package com.github.dbmdz.flusswerk.framework.engine;
 
+import static com.github.dbmdz.flusswerk.framework.fixtures.Flows.flowThrowing;
+import static com.github.dbmdz.flusswerk.framework.fixtures.Flows.passthroughFlow;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyCollection;
 import static org.mockito.Mockito.doThrow;
@@ -10,18 +12,15 @@ import static org.mockito.Mockito.when;
 
 import com.github.dbmdz.flusswerk.framework.exceptions.RetryProcessingException;
 import com.github.dbmdz.flusswerk.framework.exceptions.StopProcessingException;
+import com.github.dbmdz.flusswerk.framework.fixtures.Flows;
 import com.github.dbmdz.flusswerk.framework.flow.Flow;
-import com.github.dbmdz.flusswerk.framework.flow.builder.FlowBuilder;
 import com.github.dbmdz.flusswerk.framework.messagebroker.MessageBroker;
 import com.github.dbmdz.flusswerk.framework.model.Message;
-import com.github.dbmdz.flusswerk.framework.reporting.ReportFunction;
+import com.github.dbmdz.flusswerk.framework.reporting.SilentProcessReport;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Function;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -39,48 +38,15 @@ class EngineTest {
     message = new Message();
   }
 
-  private Flow<Message, Message, Message> passthroughFlow() {
-    return FlowBuilder.messageProcessor(Message.class).process(m -> m).build();
-  }
-
-  private Flow<Message, String, String> flowWithTransformer(Function<String, String> transformer) {
-    return FlowBuilder.flow(Message.class, String.class, String.class)
-        .reader(Message::getTracingId)
-        .transformer(transformer)
-        .writerSendingMessage(Message::new)
-        .build();
-  }
-
-  private Flow<Message, String, String> flowThrowing(Class<? extends RuntimeException> cls) {
-    var message = String.format("Generated %s for unit test", cls.getSimpleName());
-    final RuntimeException exception;
-    try {
-      exception = cls.getConstructor(String.class).newInstance(message);
-    } catch (InstantiationException
-        | IllegalAccessException
-        | InvocationTargetException
-        | NoSuchMethodException e) {
-      throw new Error("Could not instantiate exception", e); // If test is broken give up
-    }
-    Function<String, String> transformerWithException =
-        s -> {
-          throw exception;
-        };
-    return flowWithTransformer(transformerWithException);
+  private Engine createEngine(Flow<?, ?, ?> flow) {
+    return new Engine("app", messageBroker, flow, new SilentProcessReport());
   }
 
   @Test
   @DisplayName("should use the maximum number of workers")
   public void engineShouldUseMaxNumberOfWorkers() throws IOException, InterruptedException {
     when(messageBroker.receive()).thenReturn(new Message("White Room"));
-
-    Engine engine =
-        new Engine(
-            "app",
-            messageBroker,
-            flowWithTransformer(
-                new ThreadBlockingTransformer<>()) // Force engine to use all worker threads
-            );
+    Engine engine = createEngine(Flows.flowBlockingAllThreads());
     ExecutorService executorService = Executors.newSingleThreadExecutor();
     executorService.submit(engine::start);
 
@@ -105,15 +71,7 @@ class EngineTest {
   @Test
   @DisplayName("should reject a message when processing fails")
   void processShouldRejectMessageOnFailure() throws IOException {
-    Function<String, String> transformerWithException =
-        s -> {
-          throw new RuntimeException("Aaaaaaah!");
-        };
-
-    var flow = flowWithTransformer(transformerWithException);
-
-    Engine engine = new Engine("app", messageBroker, flow);
-
+    Engine engine = createEngine(flowThrowing(RuntimeException.class));
     engine.process(message);
 
     verify(messageBroker).reject(message);
@@ -123,7 +81,7 @@ class EngineTest {
   @Test
   @DisplayName("should accept a message processed without failure")
   void processShouldAcceptMessageWithoutFailure() throws IOException {
-    Engine engine = new Engine("app", messageBroker, passthroughFlow());
+    Engine engine = createEngine(passthroughFlow());
     engine.process(message);
 
     verify(messageBroker).ack(message);
@@ -133,7 +91,7 @@ class EngineTest {
   @Test
   @DisplayName("should send a message")
   void processShouldSendMessage() throws IOException {
-    Engine engine = new Engine("app", messageBroker, passthroughFlow());
+    Engine engine = createEngine(passthroughFlow());
     engine.process(new Message());
     verify(messageBroker).send(anyCollection());
   }
@@ -141,7 +99,7 @@ class EngineTest {
   @Test
   @DisplayName("should stop with retry for RetriableProcessException")
   void retryProcessExceptionShouldRejectTemporarily() throws IOException {
-    Engine engine = new Engine("app", messageBroker, flowThrowing(RetryProcessingException.class));
+    Engine engine = createEngine(flowThrowing(RetryProcessingException.class));
     engine.process(message);
 
     verify(messageBroker).reject(message);
@@ -150,28 +108,17 @@ class EngineTest {
   @Test
   @DisplayName("should stop processing for good for StopProcessingException")
   void shouldFailMessageForStopProcessingException() throws IOException {
-    Engine engine = new Engine("app", messageBroker, flowThrowing(StopProcessingException.class));
+    Engine engine = createEngine(flowThrowing(StopProcessingException.class));
     engine.process(message);
 
     verify(messageBroker).fail(message);
   }
 
   @Test
-  @DisplayName("should use functional reporter")
-  void testFunctionalReporter() {
-    final AtomicBoolean reportHasBeenCalled = new AtomicBoolean(false);
-    ReportFunction reportFn = (r, msg, e) -> reportHasBeenCalled.set(true);
-    Engine engine =
-        new Engine("app", messageBroker, flowThrowing(StopProcessingException.class), reportFn);
-    engine.process(new Message());
-    assertThat(reportHasBeenCalled.get()).isTrue();
-  }
-
-  @Test
   @DisplayName("should stop processing for good when sending messages fails")
   void shouldStopProcessingWhenSendingFails() throws IOException {
     doThrow(RuntimeException.class).when(messageBroker).send(anyCollection());
-    Engine engine = new Engine("app", messageBroker, passthroughFlow());
+    Engine engine = createEngine(passthroughFlow());
     engine.process(message);
     verify(messageBroker).fail(message);
   }
